excepciones

+++ WEB/js/app.js
@@ -2,6 +2,9 @@ $(document).ready(function() {
 
 
     $("#listar").click(function(){
+        $('#message').removeClass('alert-danger')
+        $('#message').removeClass('alert-success')
+        $('#message').text('');
        let tabla= document.querySelector('#tabla');
        tabla.innerHTML = '';
        $.ajax({
@@ -24,9 +27,17 @@ $(document).ready(function() {
 
 
     $("#buscar").click(function(){
+        $('#message').removeClass('alert-danger')
+        $('#message').removeClass('alert-success')
+        $('#message').text('');
         let tabla= document.querySelector('#tabla');
         let ref=$("#ref").val();
         tabla.innerHTML = '';
+        if(ref===''){
+            $('#message').addClass('alert-danger');
+            $('#message').text('⚠ Ingrese una referencia, para hacer la consulta');
+            return;
+        }
         $.ajax({
          url: "http://localhost:8080/producto/"+ref,
          type: "GET",
@@ -37,53 +48,98 @@ $(document).ready(function() {
              tabla.appendChild(nuevaFila);
              tabla.innerHTML+='<td>'+data.referencia+'</td><td>'+data.nombre+'</td><td>'+data.categoria+'</td><td>'+data.precioUnitario+'</td><td>'+data.cantidad+'</td><td>'+data.total  
          },
-         error: function(error){
-             console.log(error)
-         }
+         error: function(xhr, status, error) {
+            if (xhr.status === 404) {
+                $('#message').addClass('alert-danger');
+                $('#message').text('⚠ No se encontro la referencia '+ref);
+            }
+            console.log(status);
+            console.log(error);
+        }
      })
      })
 
 
      $("#eliminar").click(function(){
+        $('#message').removeClass('alert-danger')
+        $('#message').removeClass('alert-success')
+        $('#message').text('');
         let ref=$("#ref").val();
         tabla.innerHTML = '';
+        if(ref===""){
+            $('#message').addClass('alert-danger');
+            $('#message').text('⚠ Ingrese una referencia, para hacer la carga de informacion');
+            return;
+        }
         $.ajax({
          url: "http://localhost:8080/producto/eliminar/"+ref,
          type: "DELETE",
          success: function(data){ 
-            console.log("Bien hecho")
+            $('#message').addClass('alert-success');
+            $('#message').text('✅ Se elimino correctamente');
+            
          },
-         error: function(error){
-             console.log(error)
-         }
+         error: function(xhr, status, error) {
+            if (xhr.status === 404) {
+                $('#message').addClass('alert-danger');
+                $('#message').text('⚠ No se encontro la referencia '+ref);
+            }
+            console.log(status);
+            console.log(error);
+        }
      })
      })
 
      $("#cargar").click(function(){
+        $('#message').removeClass('alert-danger')
+        $('#message').removeClass('alert-success')
+        $('#message').text('');   
         let ref=$("#ref").val();
         tabla.innerHTML = '';
+        if(ref===""){
+            $('#message').addClass('alert-danger');
+            $('#message').text('⚠ Ingrese una referencia, para hacer la carga de informacion');
+            return;
+        }
         $.ajax({
          url: "http://localhost:8080/producto/"+ref,
          type: "GET",
          datatype: "JSON",
          success: function(data){
+            $('#actualizar').prop('disabled', false)
+            $('#referencia').prop('disabled', true);
+            $('#agregar').prop('disabled', true);
             $("#referencia").val(data.referencia);
             $("#nombre").val(data.nombre); 
             $("#categoria").val(data.categoria); 
             $("#precioUnitario").val(data.precioUnitario); 
             $("#cantidad").val(data.cantidad); 
+            $('#message').addClass('alert-success');
+            $('#message').text('✅ Se cargo correctamente la informacion, por favor actualiza.');
          },
-         error: function(error){
-             console.log(error)
-         }
+         error: function(xhr, status, error) {
+            if (xhr.status === 404) {
+                $('#message').addClass('alert-danger');
+                $('#message').text('⚠ No se encontro la referencia '+ref);
+            }
+            console.log(status);
+            console.log(error);
+        }
      })
      })
      $("#agregar").click(function(){
+        $('#mensaje').removeClass('alert-danger')
+        $('#mensaje').removeClass('alert-success')
+        $('#mensaje').text('');   
         let referencia=$("#referencia").val();
         let nombre=$("#nombre").val();
         let categoria=$("#categoria").val();
         let precioUnitario=$("#precioUnitario").val();
         let cantidad=$("#cantidad").val();
+        if(referencia==='' || nombre==='' || categoria==='' || cantidad==='' || precioUnitario===''){
+            $('#mensaje').addClass('alert-danger');
+            $('#mensaje').text('⚠ Completa todos los campos');
+        }
         let datos={
             referencia: referencia, 
             nombre:nombre,
@@ -97,20 +153,34 @@ $(document).ready(function() {
          data: JSON.stringify(datos),
          contentType: "application/json",
          success: function(data){
-
+            $('#form')[0].reset();
+            $('#mensaje').addClass('alert-success');
+            $('#mensaje').text('✅ Se guardo correctamente');
          },
-         error: function(error){
-             console.log(error)
-         }
+         error: function(xhr, status, error) {
+            if (xhr.status === 403) {
+                $('#mensaje').addClass('alert-danger');
+                $('#mensaje').text('⚠ Ya existe la referencia'+referencia);
+            }
+            console.log(status);
+            console.log(error);
+        }
      })
      })
 
      $("#actualizar").click(function(){
+        $('#mensaje').removeClass('alert-danger')
+        $('#mensaje').removeClass('alert-success')
+        $('#mensaje').text('');  
         let referencia=$("#referencia").val();
         let nombre=$("#nombre").val();
         let categoria=$("#categoria").val();
         let precioUnitario=$("#precioUnitario").val();
         let cantidad=$("#cantidad").val();
+        if(referencia==='' || nombre==='' || categoria==='' || cantidad==='' || precioUnitario===''){
+            $('#mensaje').addClass('alert-danger');
+            $('#mensaje').text('⚠ Completa todos los campos');
+        }
         let datos={
             referencia: referencia, 
             nombre:nombre,
@@ -124,25 +194,51 @@ $(document).ready(function() {
          data: JSON.stringify(datos),
          contentType: "application/json",
          success: function(data){
+            $('#form')[0].reset();
+            $('#mensaje').addClass('alert-success');
+            $('#mensaje').text('✅ Se actualizo correctamente');
+            $('#actualizar').prop('disabled', true)
+            $('#referencia').prop('disabled', false);
+            $('#agregar').prop('disabled', false);
+
          },
-         error: function(error){
-             console.log(error)
-         }
+         error: function(xhr, status, error) {
+            console.log(status);
+            console.log(error);
+        }
      })
      })
 
 
      $("#actualizarCantidad").click(function(){
+        $('#mensajeActu').removeClass('alert-danger')
+        $('#mensajeActu').removeClass('alert-success')
+        $('#mensajeActu').text('');   
         let categoria=$("#porCategoria").val();
         let cantidad=$("#porCantidad").val();
+        if(categoria==='' || cantidad===''){
+            $('#mensajeActu').addClass('alert-danger');
+            $('#mensajeActu').text('⚠ Completa todos los campos');
+        }
         $.ajax({
          url: "http://localhost:8080/producto/actualizarCantidad/"+categoria+'/'+cantidad,
          type: "PUT",
          success: function(data){
+            $('#mensajeActu').addClass('alert-success');
+            $('#mensajeActu').text('✅ Se actualizo la cantidad correctamente');
          },
-         error: function(error){
-             console.log(error)
-         }
+         error: function(xhr, status, error) {
+            if (xhr.status === 409) {
+                $('#mensajeActu').addClass('alert-danger');
+                $('#mensajeActu').text('⚠ La cantidad debe ser un numero mayor o igual a 0');
+            }else if (xhr.status ===404){
+                $('#mensajeActu').addClass('alert-danger');
+                $('#mensajeActu').text('⚠ La categoria no existe.');
+
+            }
+            console.log(status);
+            console.log(error);
+        }
      })
      })
